+++ NEWS.md
@@ -169,6 +169,8 @@ mu-referencing style to run the optimization.
   `ini(param=NULL)` changes the parameter to a covariate to align with
   this idiom of dropping parameters
 
+- `rxRename` has been refactored to run faster
+
 ## Internal new features
 
 - Add `as.model()` for list expressions, which implies `model(ui) <-
@@ -187,6 +189,9 @@ mu-referencing style to run the optimization.
   version from Morwenn
 
 ## Bug fixes
+
+- Simulating/solving from functions/ui now prefers params over `omega`
+  and `sigma` in the model (#632)
 
 - Piping does not add constants to the initial estimates